refactor: extract Max HR calculation to shared utility function
The Max HR calculation logic (220 - age) and the default Max HR constant (185) have been extracted from `lib/shared/hr-zones.ts` into a new dedicated utility file `utils/hrCalculations.ts`.

Key changes:
- Created `utils/hrCalculations.ts` to encapsulate `calculateMaxHr` and `MAX_HR_DEFAULT`.
- Updated `lib/shared/hr-zones.ts` to re-export these from the new utility to maintain backward compatibility.
- Updated all import sites across the application (ExperimentalAnalyticsPage, mock/page, connect/page, useBluetoothHRM, useWorkoutSessionManager, hrm util) and their corresponding tests to use the new utility import.
- Fixed minor linting issues discovered during the refactoring.

This centralization improves code consistency and facilitates future updates to the Max HR formula.
