+++ hooks/useBluetoothHRM.ts
@@ -5,7 +5,7 @@ import {
 } from '@/types/websocket'
 import { BluetoothConnectionStatus } from '@/types/bluetooth'
 import isEqual from 'lodash.isequal'
-import { calculateMaxHr } from '@/lib/shared/hr-zones'
+import { calculateMaxHr } from '@/utils/hrCalculations'
 import logger from '@/utils/logger'
 import { useWebSocket } from '@/context/WebSocketContext'
 import { cancellablePromise } from '@/utils/promise'
