+++ utils/hrCalculations.ts
@@ -1,0 +1,27 @@
+/**
+ * Shared utility functions for heart rate calculations.
+ */
+
+export const MAX_HR_DEFAULT = 185
+
+/**
+ * Estimates a user's maximum heart rate using the Haskell & Fox formula (220 - age).
+ *
+ * NOTE: While the Tanaka formula (208 - 0.7 * age) is often more accurate for older adults,
+ * we are using the Haskell & Fox formula here to ensure parity with existing dashboard
+ * calculation logic and requested test cases (e.g. 120-year-old athlete).
+ *
+ * @param age - The user's age in years.
+ * @returns The estimated maximum heart rate.
+ */
+export const calculateMaxHr = (age?: number | string | null): number => {
+  if (!age) return MAX_HR_DEFAULT
+
+  const ageNum = typeof age === 'string' ? parseInt(age, 10) : age
+
+  if (isNaN(ageNum) || ageNum <= 0) {
+    return MAX_HR_DEFAULT
+  }
+
+  return 220 - ageNum
+}
